- 新增： shardingjdbc 日志分表 - 优化： 多角色权限问题

+++ pig-modules/pig-upms-service/src/main/java/com/github/pig/admin/controller/LogController.java
@@ -46,7 +46,7 @@ public class LogController extends BaseController {
      * @return success/false
      */
     @DeleteMapping("/{id}")
-    public R<Boolean> delete(@PathVariable Integer id) {
+    public R<Boolean> delete(@PathVariable Long id) {
         return new R<>(sysLogService.updateByLogId(id));
     }
 }
